refactor(gui): split draw_page into submods; rename slot→area; area width tracks page
- draw_page.rs split into draw_page/{hud,overlays,manual}.rs
- draw_new_page_slot.rs renamed to draw_new_page_area.rs; draw() split
  into classify_drag / draw_zone_background / draw_zone_label submethods
- drop zone width now matches adjacent page width instead of fixed 520px
- draw_pages.rs passes page_rect.width() to each area draw call
