+++ gui/app/widgets/central_panel/draw_page.rs
@@ -1,12 +1,14 @@
-use crate::state::{ActiveDrag, DataState, DragMode, DragSource, InteractionState, PageHudAnim};
+mod hud;
+mod manual;
+mod overlays;
+
+use crate::state::{DataState, InteractionState, PageHudAnim};
 use crate::task::BackgroundTask;
 
-use super::super::geometry::{self, PageDimensions};
-use super::theme::FbTheme;
+use super::super::geometry::PageDimensions;
 use super::{draw_drag_ghosts, helpers};
 
-const HUD_HEIGHT: f32 = 24.0;
-const HUD_GAP: f32 = 10.0;
+use hud::{HUD_GAP, HUD_HEIGHT};
 
 /// Returns `(hovered_slot, over_page, page_rect, cursor_mm)`.
 pub(super) fn draw_page(
@@ -26,14 +28,9 @@ pub(super) fn draw_page(
     };
     let page_size = helpers::page_display_size(interaction.viewport.zoom, dims);
 
-    // Render the page image through the normal egui layout so cursor advances
-    // exactly by page_size — no pre-allocation block that interferes later.
     let page_rect = render_page_image(ui, data, page_idx, page_size);
-
-    // Reserve space for the gap + HUD strip below the page.
     ui.add_space(HUD_GAP + HUD_HEIGHT);
 
-    // Hover over the combined block: page rect + HUD area below it.
     let block_bottom = page_rect.max.y + HUD_GAP + HUD_HEIGHT;
     let block_rect =
         egui::Rect::from_min_max(page_rect.min, egui::pos2(page_rect.max.x, block_bottom));
@@ -42,242 +39,69 @@ pub(super) fn draw_page(
         .input(|i| i.pointer.latest_pos().map(|p| block_rect.contains(p)))
         .unwrap_or(false);
 
-    // Advance animation.
     let dt = ui.ctx().input(|i| i.unstable_dt).min(0.05);
     let anim = interaction
         .page_hud
         .entry(page_idx)
         .or_insert_with(PageHudAnim::default);
-    let still_moving = anim.advance(hovered, dt);
-    if still_moving {
+    if anim.advance(hovered, dt) {
         ui.ctx().request_repaint();
     }
 
-    if let Some(layout_page) = data.project.layout.get(page_idx) {
-        draw_slot_overlays(ui, page_rect, data, interaction, page_idx, dims);
-        let (hovered_slot, over_page, cursor_mm) =
-            hit_test_pointer(ui, page_rect, layout_page, dims);
-        super::super::page_nav::draw_nav_selection_overlay(ui, interaction, page_idx, page_rect);
-        draw_page_move_highlight(ui, interaction, page_idx, page_rect, over_page);
-        draw_pool_drag_overlay(ui, interaction, page_idx, page_rect);
-        draw_drag_ghosts::draw_drag_ghosts(ui, data, interaction, page_idx, page_rect, dims);
-
-        use fotobuch::dto_models::PageMode;
-        if layout_page.mode == PageMode::Manual {
-            let full_w_mm = dims.width_mm + 2.0 * dims.bleed_mm;
-            let pixel_per_mm = if full_w_mm > 0.0 {
-                page_rect.width() as f64 / full_w_mm
-            } else {
-                1.0
-            };
-            draw_manual_handles_and_overlay(
-                ui,
-                data,
-                interaction,
-                page_idx,
-                page_rect,
-                dims,
-                pixel_per_mm,
-            );
-        }
-
-        // Snapshot anim values (borrow ends after this block).
-        let (opacity, pill_w, actions_alpha, actions_offset) = {
-            let a = &interaction.page_hud[&page_idx];
-            (a.opacity, a.pill_width, a.actions_alpha, a.actions_offset)
-        };
-
-        let hud_rect = egui::Rect::from_min_size(
-            egui::pos2(page_rect.min.x, page_rect.max.y + HUD_GAP),
-            egui::vec2(page_rect.width(), HUD_HEIGHT),
-        );
-        draw_hud(
-            ui,
-            hud_rect,
-            page_idx,
-            layout_page.mode,
-            opacity,
-            pill_w,
-            actions_alpha,
-            actions_offset,
-            cmds,
-        );
-
-        (hovered_slot, over_page, page_rect, cursor_mm)
-    } else {
-        (None, false, page_rect, (0.0, 0.0))
-    }
-}
-
-#[allow(clippy::too_many_arguments)]
-fn draw_hud(
-    ui: &mut egui::Ui,
-    hud_rect: egui::Rect,
-    page_idx: usize,
-    mode: fotobuch::dto_models::PageMode,
-    opacity: f32,
-    pill_width: f32,
-    actions_alpha: f32,
-    actions_offset: f32,
-    cmds: &mut Vec<BackgroundTask>,
-) {
-    use fotobuch::dto_models::PageMode;
-
-    let center_y = hud_rect.center().y;
-
-    // The dot/pill is anchored at the horizontal center of the HUD (below page center).
-    // The label grows leftward from it; actions grow rightward — so the dot never jumps.
-    let label_w: f32 = 28.0;
-    let btn_size: f32 = 22.0;
-    let gap: f32 = 10.0;
-    let dot_center_x = hud_rect.center().x;
-
-    let alpha_u8 = (opacity * 255.0).clamp(0.0, 255.0) as u8;
-
-    // Page number label — always gap+half-pill to the left of dot center.
-    let label_str = format!("{}", page_idx + 1);
-    let label_center = egui::pos2(
-        dot_center_x - pill_width / 2.0 - gap - label_w / 2.0,
-        center_y,
-    );
-    ui.painter().text(
-        label_center,
-        egui::Align2::CENTER_CENTER,
-        &label_str,
-        egui::FontId::monospace(10.5),
-        FbTheme::with_alpha(FbTheme::TEXT_MUTE, alpha_u8),
-    );
-
-    // Mode dot / pill — centered on dot_center_x at all animation stages.
-    let mode_color = match mode {
-        PageMode::Auto => FbTheme::AUTO,
-        PageMode::Manual => FbTheme::MANUAL,
+    let Some(layout_page) = data.project.layout.get(page_idx) else {
+        return (None, false, page_rect, (0.0, 0.0));
     };
 
-    // Height tracks width up to 18 px so the shape is always a circle at rest.
-    // corner_radius is clamped to half the smaller dimension so egui never clips
-    // the corners into an oval/diamond artefact when width < 2*radius.
-    let pill_height = pill_width.min(18.0);
-    let corner_radius = (pill_width / 2.0).min(pill_height / 2.0);
-    let pill_rect = egui::Rect::from_center_size(
-        egui::pos2(dot_center_x, center_y),
-        egui::vec2(pill_width, pill_height),
-    );
-    // 0.0 = idle circle (10 px), 1.0 = fully open pill (80 px).
-    let expand_frac = ((pill_width - 10.0) / (80.0 - 10.0)).clamp(0.0, 1.0);
+    overlays::draw_slot_overlays(ui, page_rect, data, interaction, page_idx, dims);
 
-    let lerp_u8 = |a: u8, b: u8, t: f32| (a as f32 + (b as f32 - a as f32) * t) as u8;
+    let (hovered_slot, over_page, cursor_mm) = hit_test_pointer(ui, page_rect, layout_page, dims);
 
-    // Fill: solid gray circle → translucent mode-colored pill.
-    let fill_r = lerp_u8(FbTheme::TEXT_MUTE.r(), mode_color.r(), expand_frac);
-    let fill_g = lerp_u8(FbTheme::TEXT_MUTE.g(), mode_color.g(), expand_frac);
-    let fill_b = lerp_u8(FbTheme::TEXT_MUTE.b(), mode_color.b(), expand_frac);
-    let fill_a = lerp_u8(alpha_u8, (alpha_u8 as f32 * 0.13) as u8, expand_frac);
-    let fill_color = egui::Color32::from_rgba_unmultiplied(fill_r, fill_g, fill_b, fill_a);
+    super::super::page_nav::draw_nav_selection_overlay(ui, interaction, page_idx, page_rect);
+    overlays::draw_page_move_highlight(ui, interaction, page_idx, page_rect, over_page);
+    overlays::draw_pool_drag_overlay(ui, interaction, page_idx, page_rect);
+    draw_drag_ghosts::draw_drag_ghosts(ui, data, interaction, page_idx, page_rect, dims);
 
-    // Border: invisible at circle, mode-colored at pill.
-    let stroke_a = (alpha_u8 as f32 * 0.40 * expand_frac) as u8;
-    let stroke_color = egui::Color32::from_rgba_unmultiplied(
-        mode_color.r(),
-        mode_color.g(),
-        mode_color.b(),
-        stroke_a,
-    );
-
-    ui.painter().rect(
-        pill_rect,
-        corner_radius,
-        fill_color,
-        egui::Stroke::new(1.0, stroke_color),
-        egui::StrokeKind::Inside,
-    );
-
-    // Text fades in as pill expands.
-    if expand_frac > 0.01 {
-        let pill_label = match mode {
-            PageMode::Auto => "✦ AUTO",
-            PageMode::Manual => "✋ MANUAL",
+    use fotobuch::dto_models::PageMode;
+    if layout_page.mode == PageMode::Manual {
+        let full_w_mm = dims.width_mm + 2.0 * dims.bleed_mm;
+        let pixel_per_mm = if full_w_mm > 0.0 {
+            page_rect.width() as f64 / full_w_mm
+        } else {
+            1.0
         };
-        ui.painter().text(
-            pill_rect.center(),
-            egui::Align2::CENTER_CENTER,
-            pill_label,
-            egui::FontId::monospace(10.0),
-            FbTheme::with_alpha(mode_color, (expand_frac * alpha_u8 as f32) as u8),
+        manual::draw_manual_handles_and_overlay(
+            ui,
+            data,
+            interaction,
+            page_idx,
+            page_rect,
+            dims,
+            pixel_per_mm,
         );
     }
 
-    // Click detection via raw input — avoids ui.allocate_rect which expands
-    // min_rect and causes layout shifts when the pill changes size.
-    let pointer = ui
-        .ctx()
-        .input(|i| i.pointer.interact_pos().filter(|_| i.pointer.any_click()));
-    if pointer.is_some_and(|p| pill_rect.contains(p)) {
-        let new_mode = match mode {
-            PageMode::Auto => PageMode::Manual,
-            PageMode::Manual => PageMode::Auto,
-        };
-        cmds.push(BackgroundTask::SetPageMode {
-            page: page_idx,
-            mode: new_mode,
-        });
-    }
+    let (opacity, pill_w, actions_alpha, actions_offset) = {
+        let a = &interaction.page_hud[&page_idx];
+        (a.opacity, a.pill_width, a.actions_alpha, a.actions_offset)
+    };
 
-    // Action buttons (fade in with actions_alpha)
-    if actions_alpha > 0.001 {
-        let act_alpha = (actions_alpha * alpha_u8 as f32) as u8;
-        let btn_x = dot_center_x + pill_width / 2.0 + gap + actions_offset;
+    let hud_rect = egui::Rect::from_min_size(
+        egui::pos2(page_rect.min.x, page_rect.max.y + HUD_GAP),
+        egui::vec2(page_rect.width(), HUD_HEIGHT),
+    );
+    hud::draw_hud(
+        ui,
+        hud_rect,
+        page_idx,
+        layout_page.mode,
+        opacity,
+        pill_w,
+        actions_alpha,
+        actions_offset,
+        cmds,
+    );
 
-        // ↻ Rebuild button
-        let rebuild_rect = egui::Rect::from_min_size(
-            egui::pos2(btn_x, center_y - btn_size / 2.0),
-            egui::vec2(btn_size, btn_size),
-        );
-        ui.painter().rect(
-            rebuild_rect,
-            4.0,
-            egui::Color32::TRANSPARENT,
-            egui::Stroke::new(1.0, FbTheme::with_alpha(FbTheme::STROKE, act_alpha)),
-            egui::StrokeKind::Inside,
-        );
-        ui.painter().text(
-            rebuild_rect.center(),
-            egui::Align2::CENTER_CENTER,
-            "↻",
-            egui::FontId::proportional(14.0),
-            FbTheme::with_alpha(FbTheme::TEXT_DIM, act_alpha),
-        );
-        if pointer.is_some_and(|p| rebuild_rect.contains(p)) {
-            cmds.push(BackgroundTask::RebuildPages {
-                pages: vec![page_idx],
-            });
-        }
-
-        // ✕ Delete button
-        let delete_rect = egui::Rect::from_min_size(
-            egui::pos2(btn_x + btn_size + gap, center_y - btn_size / 2.0),
-            egui::vec2(btn_size, btn_size),
-        );
-        ui.painter().rect(
-            delete_rect,
-            4.0,
-            egui::Color32::TRANSPARENT,
-            egui::Stroke::new(1.0, FbTheme::with_alpha(FbTheme::STROKE, act_alpha)),
-            egui::StrokeKind::Inside,
-        );
-        ui.painter().text(
-            delete_rect.center(),
-            egui::Align2::CENTER_CENTER,
-            "✕",
-            egui::FontId::proportional(12.0),
-            FbTheme::with_alpha(FbTheme::DANGER, act_alpha),
-        );
-        if pointer.is_some_and(|p| delete_rect.contains(p)) {
-            cmds.push(BackgroundTask::DeletePages {
-                pages: vec![page_idx],
-            });
-        }
-    }
+    (hovered_slot, over_page, page_rect, cursor_mm)
 }
 
 fn render_page_image(
@@ -314,100 +138,13 @@ fn render_page_image(
     rect
 }
 
-fn draw_slot_overlays(
-    ui: &mut egui::Ui,
-    page_rect: egui::Rect,
-    data: &DataState,
-    interaction: &InteractionState,
-    page_idx: usize,
-    dims: PageDimensions,
-) {
-    let layout_page = match data.project.layout.get(page_idx) {
-        Some(lp) => lp,
-        None => return,
-    };
-
-    let is_slot_drag = matches!(
-        interaction.drag.active,
-        ActiveDrag::Dragging(DragSource::Slot { .. })
-    );
-    let is_swap_drag = is_slot_drag && interaction.drag.mode == DragMode::Swap;
-
-    use fotobuch::dto_models::PageMode;
-    let src_is_manual =
-        if let ActiveDrag::Dragging(DragSource::Slot { src_page, .. }) = &interaction.drag.active {
-            data.project
-                .layout
-                .get(*src_page)
-                .map(|p| p.mode == PageMode::Manual)
-                .unwrap_or(false)
-        } else {
-            false
-        };
-    let this_is_manual = data
-        .project
-        .layout
-        .get(page_idx)
-        .map(|p| p.mode == PageMode::Manual)
-        .unwrap_or(false);
-    let is_swap_drag = is_swap_drag && !src_is_manual && !this_is_manual;
-
-    let drag_src_ratio: Option<f64> =
-        if let ActiveDrag::Dragging(DragSource::Slot {
-            src_page, src_slot, ..
-        }) = &interaction.drag.active
-        {
-            data.project
-                .layout
-                .get(*src_page)
-                .and_then(|p| p.slots.get(*src_slot))
-                .map(|s| s.width_mm / s.height_mm)
-        } else {
-            None
-        };
-
-    let painter = ui.painter();
-    let pointer_pos = ui.input(|i| i.pointer.latest_pos());
-    for (slot_idx, slot) in layout_page.slots.iter().enumerate() {
-        let slot_rect = geometry::slot_rect_on_screen(page_rect, dims, slot);
-        let is_hovered = pointer_pos.map(|p| slot_rect.contains(p)).unwrap_or(false);
-
-        if is_swap_drag {
-            let target_ratio = slot.width_mm / slot.height_mm;
-            let same_ratio =
-                drag_src_ratio.is_some_and(|r| geometry::slot_ratio_similar(r, target_ratio));
-            let alpha = if is_hovered { 220 } else { 140 };
-            let color = if same_ratio {
-                egui::Color32::from_rgba_unmultiplied(0, 200, 80, alpha)
-            } else {
-                egui::Color32::from_rgba_unmultiplied(220, 50, 50, alpha)
-            };
-            painter.rect_filled(slot_rect, 0.0, color);
-        } else if is_hovered && !is_slot_drag {
-            painter.rect_filled(
-                slot_rect,
-                0.0,
-                egui::Color32::from_rgba_unmultiplied(0, 120, 255, 38),
-            );
-        }
-
-        if interaction.selections.slots.is_selected(page_idx, slot_idx) {
-            painter.rect_stroke(
-                slot_rect,
-                0.0,
-                egui::Stroke::new(2.0, egui::Color32::from_rgb(50, 200, 80)),
-                egui::StrokeKind::Middle,
-            );
-        }
-    }
-}
-
 fn hit_test_pointer(
     ui: &mut egui::Ui,
     page_rect: egui::Rect,
     layout_page: &fotobuch::dto_models::LayoutPage,
     dims: PageDimensions,
 ) -> (Option<usize>, bool, (f32, f32)) {
+    use super::super::geometry;
     match ui.input(|i| i.pointer.latest_pos()) {
         None => (None, false, (0.0, 0.0)),
         Some(pos) => {
@@ -423,189 +160,5 @@ fn hit_test_pointer(
             };
             (slot, over, cursor_mm)
         }
-    }
-}
-
-fn draw_page_move_highlight(
-    ui: &mut egui::Ui,
-    interaction: &InteractionState,
-    page_idx: usize,
-    page_rect: egui::Rect,
-    over_page: bool,
-) {
-    let is_move_drag = matches!(
-        interaction.drag.active,
-        ActiveDrag::Dragging(DragSource::Slot { .. })
-    ) && interaction.drag.mode == DragMode::Move;
-    if !is_move_drag || !over_page {
-        return;
-    }
-    let is_src_page = matches!(
-        interaction.drag.active,
-        ActiveDrag::Dragging(DragSource::Slot { src_page, .. }) if src_page == page_idx
-    );
-    if is_src_page {
-        return;
-    }
-    ui.painter().rect_stroke(
-        page_rect,
-        0.0,
-        egui::Stroke::new(3.0, egui::Color32::from_rgba_unmultiplied(0, 200, 80, 180)),
-        egui::StrokeKind::Inside,
-    );
-}
-
-fn draw_pool_drag_overlay(
-    ui: &mut egui::Ui,
-    interaction: &InteractionState,
-    page_idx: usize,
-    page_rect: egui::Rect,
-) {
-    if matches!(
-        interaction.drag.active,
-        ActiveDrag::Dragging(DragSource::Pool { .. })
-    ) && interaction.hovered.as_ref().and_then(|h| h.central_page()) == Some(page_idx)
-    {
-        ui.painter().rect_filled(
-            page_rect,
-            0.0,
-            egui::Color32::from_rgba_unmultiplied(64, 128, 255, 48),
-        );
-    }
-}
-
-fn se_corner_rect(slot_rect: egui::Rect) -> egui::Rect {
-    const SZ: f32 = 8.0;
-    egui::Rect::from_center_size(slot_rect.right_bottom(), egui::vec2(SZ, SZ))
-}
-
-#[allow(clippy::too_many_arguments)]
-fn draw_manual_handles_and_overlay(
-    ui: &mut egui::Ui,
-    data: &DataState,
-    interaction: &mut InteractionState,
-    page_idx: usize,
-    page_rect: egui::Rect,
-    dims: PageDimensions,
-    pixel_per_mm: f64,
-) {
-    let layout_page = match data.project.layout.get(page_idx) {
-        Some(lp) => lp,
-        None => return,
-    };
-
-    let cursor = ui.input(|i| i.pointer.hover_pos()).unwrap_or_default();
-    let rmb_pressed = ui.input(|i| i.pointer.secondary_pressed());
-
-    if rmb_pressed
-        && interaction.drag.mode == DragMode::Move
-        && matches!(interaction.drag.active, ActiveDrag::Idle)
-    {
-        for (slot_idx, slot) in layout_page.slots.iter().enumerate().rev() {
-            let slot_rect = geometry::slot_rect_on_screen(page_rect, dims, slot);
-            let se = se_corner_rect(slot_rect);
-            let source = if se.contains(cursor) {
-                Some(DragSource::ManualResize {
-                    page: page_idx,
-                    slot: slot_idx,
-                    pointer_origin: cursor,
-                    slot_origin_mm: (slot.x_mm, slot.y_mm, slot.width_mm, slot.height_mm),
-                    pixel_per_mm,
-                })
-            } else if slot_rect.contains(cursor) {
-                Some(DragSource::ManualMove {
-                    page: page_idx,
-                    slot: slot_idx,
-                    pointer_origin: cursor,
-                    slot_origin_mm: (slot.x_mm, slot.y_mm),
-                    pixel_per_mm,
-                })
-            } else {
-                None
-            };
-            if let Some(src) = source {
-                interaction.drag.active = ActiveDrag::Pending {
-                    source: src,
-                    press_pos: cursor,
-                    press_instant: std::time::Instant::now(),
-                };
-                break;
-            }
-        }
-    }
-
-    let manual_dragging = matches!(
-        &interaction.drag.active,
-        ActiveDrag::Dragging(DragSource::ManualMove { page, .. } | DragSource::ManualResize { page, .. })
-            if *page == page_idx
-    );
-    if !manual_dragging {
-        for slot in &layout_page.slots {
-            let slot_rect = geometry::slot_rect_on_screen(page_rect, dims, slot);
-            let se = se_corner_rect(slot_rect);
-            ui.painter().rect_filled(
-                se,
-                0.0,
-                egui::Color32::from_rgba_unmultiplied(255, 200, 0, 200),
-            );
-        }
-    }
-
-    match &interaction.drag.active {
-        ActiveDrag::Dragging(DragSource::ManualMove {
-            page,
-            slot,
-            pointer_origin,
-            slot_origin_mm,
-            pixel_per_mm: ppm,
-        }) if *page == page_idx => {
-            let delta_px = cursor - *pointer_origin;
-            let dx_mm = delta_px.x as f64 / ppm;
-            let dy_mm = delta_px.y as f64 / ppm;
-            if let Some(slot_data) = layout_page.slots.get(*slot) {
-                let preview = fotobuch::dto_models::Slot {
-                    x_mm: slot_origin_mm.0 + dx_mm,
-                    y_mm: slot_origin_mm.1 + dy_mm,
-                    width_mm: slot_data.width_mm,
-                    height_mm: slot_data.height_mm,
-                };
-                let r = geometry::slot_rect_on_screen(page_rect, dims, &preview);
-                ui.painter().rect_stroke(
-                    r,
-                    0.0,
-                    egui::Stroke::new(2.0, egui::Color32::from_rgb(255, 128, 0)),
-                    egui::StrokeKind::Outside,
-                );
-            }
-        }
-        ActiveDrag::Dragging(DragSource::ManualResize {
-            page,
-            slot,
-            pointer_origin,
-            slot_origin_mm,
-            pixel_per_mm: ppm,
-        }) if *page == page_idx => {
-            let delta_px = cursor - *pointer_origin;
-            let (_, _, new_w, new_h) =
-                super::manual_resize::compute_se(*slot_origin_mm, delta_px, *ppm);
-            if let Some(slot_data) = layout_page.slots.get(*slot) {
-                let preview = fotobuch::dto_models::Slot {
-                    x_mm: slot_origin_mm.0,
-                    y_mm: slot_origin_mm.1,
-                    width_mm: new_w,
-                    height_mm: new_h,
-                };
-                let r = geometry::slot_rect_on_screen(page_rect, dims, &preview);
-                ui.painter().rect_stroke(
-                    r,
-                    0.0,
-                    egui::Stroke::new(2.0, egui::Color32::from_rgb(0, 200, 255)),
-                    egui::StrokeKind::Outside,
-                );
-                let _ = slot_data;
-            }
-            let _ = new_h;
-        }
-        _ => {}
     }
 }
